test(e2e): address review — isolate restart in stateful/, harden filter assert, fix dialog testid
- Move the real daemon-restart test out of the top-level pages.spec.ts into
  tests/admin-app/stateful/restart.spec.ts, per the suite README's
  "Spec ordering (stateful/)" rule. Its path now sorts after every read-only
  admin-app spec (pages/pwa-shell/smoke), so a slow or flaky restart can't
  cascade into them — matching how admin-site isolates power.spec.

- Make the device-filter assertion race-free: re-read the pill count and row
  count together each poll via expect.poll, instead of snapshotting the
  time-derived "online" count into a fixed toHaveCount target that a refetch
  could leave unsatisfiable.

- Move the confirm-dialog testid from the inner buttons div onto DrawerContent
  (the dialog root; it forwards data-testid), so the testid actually scopes the
  whole dialog (title + buttons), not just the button column.

- Drop the unused device-routing-tunnel-${id} testid; the spec selects the
  tunnel option by accessible name.
